fix: fix minor bugs and typo

+++ backend/src/controllers/postController.js
@@ -25,9 +25,10 @@ async function getPost(req, res, next) {
 
 async function getPostComments(req, res, next) {
   try {
+		const postId = Number(req.params.postId)
     const comments = await prisma.$queryRaw`
       SELECT comment_content, "User".user_name FROM "Comment" INNER JOIN "User" ON
-      "Comment"."author_id"="User"."user_id";
+      "Comment"."author_id"="User"."user_id" WHERE "Post".post_id = ${postId};
     `;
     return res.status(200).json({success: true, comments})
   } catch(err) {
@@ -113,8 +114,13 @@ async function deletePost(req, res, next) {
         post_id: Number(req.params.postId)
       }
     })
+
 		if (!post) {
 			return res.status(404).json({success: false, message: "Post not found"})
+		}
+
+		if (!user.author && user.user_id != post.author_id) {
+			return res.status(401).json({success: false, message: "Unauthorized!"})
 		}
 
 		const delPost = await prisma.post.delete({
@@ -149,10 +155,10 @@ async function deletePostComment(req, res, next) {
     });
 
 		if (!comment) {
-			return res.statsu(404).json({success: false, message: "Comment not found!"})
+			return res.status(404).json({success: false, message: "Comment not found!"})
 		}
 
-		if (comment.author_id != req.user.userid || !user.admin) {
+		if (comment.author_id != req.user.userid && !user.admin) {
 			return res.status(401).json({success: false, message: "Unauthorized!"})
 		}
 
@@ -180,15 +186,15 @@ async function updatePost(req, res, next) {
 			}
 		})
 
-		if (!user.author) {
-			return res.status(401).json({success: false, message: "Unauthorized!"})
-		}
-
     const post = await prisma.post.findUnique({
       where: {
         post_id: Number(req.params.postId)
       }
     })
+
+		if (!user.author && user.user_id != post.author_id) {
+			return res.status(401).json({success: false, message: "Unauthorized!"})
+		}
 
 		if (!post) {
 			return res.status(404).json({success: false, message: "Post not found!"})
@@ -224,7 +230,7 @@ async function publishPost(req, res, next) {
 
 		const post = await prisma.post.findUnique({
 			where: {
-				id: Number(req.params.postId)
+				post_id: Number(req.params.postId)
 			}
 		})
 
@@ -269,7 +275,7 @@ async function unpublishPost(req, res, next) {
 
 		const post = await prisma.post.findUnique({
 			where: {
-				id: Number(req.params.postId)
+				post_id: Number(req.params.postId)
 			}
 		})
